+++ src/main/java/org/bytedeco/javacpp/PointerPointer.java
@@ -129,16 +129,16 @@ public class PointerPointer<P extends Pointer> extends Pointer {
     /** This is just to keep references to Pointer objects and prevent premature deallocation. */
     private P[] pointerArray;
 
-    /** @see Pointer#position(int) */
-    @Override public PointerPointer<P> position(int position) {
+    /** @see Pointer#position(long) */
+    @Override public PointerPointer<P> position(long position) {
         return super.position(position);
     }
-    /** @see Pointer#limit(int) */
-    @Override public PointerPointer<P> limit(int limit) {
+    /** @see Pointer#limit(long) */
+    @Override public PointerPointer<P> limit(long limit) {
         return super.limit(limit);
     }
-    /** @see Pointer#capacity(int) */
-    @Override public PointerPointer<P> capacity(int capacity) {
+    /** @see Pointer#capacity(long) */
+    @Override public PointerPointer<P> capacity(long capacity) {
         return super.capacity(capacity);
     }
